sinBasis_main_code/requirements.txt
numpy
scipy
torchmetrics>=1.3.0
tqdm
pillow
opencv-python
einops
matplotlib
tensorboard
omegaconf>=2.3.0
hydra-core>=1.3.2
rich
timm>=1.0.8
h5py
